library/requirements.txt
click
ollama
pillow
pylette
python-dotenv
supabase
tqdm